Added a new folder for a new practice exercise: 'Web Scraping Exercise', containing output files and script.
Created a new folder named 'Web Scraping Exercise' to house all files related to the web scraping exercise. Inside this folder, an 'Output' subfolder was added, which contains the generated files 'Movies.db' and 'top_50_films.csv'. Additionally, a 'Script' subfolder was created to hold the script 'webscraping_movies.py', which contains the code for the exercise.