Updated main and Final Project README, added final project code, database, log, and CSV files
Updated the main README.md file to provide a better visualization and representation of the project. Made improvements to the Final Project README.md for clearer instructions and guidance.

Added the code for the final project, along with the database file, log file, and CSV file that were created as a result of running the final project code.